// openmetadata-ui/src/main/resources/ui/src/enums/entity.enum.ts
export enum EntityType {
  TABLE = 'table',
  TOPIC = 'topic',
  CLASSIFICATION = 'classification',
  DASHBOARD = 'dashboard',
  PIPELINE = 'pipeline',
  DATABASE = 'database',
  DATABASE_SCHEMA = 'databaseSchema',
  GLOSSARY = 'glossary',
  GLOSSARY_TERM = 'glossaryTerm',
  DATABASE_SERVICE = 'databaseService',
  MESSAGING_SERVICE = 'messagingService',
  METADATA_SERVICE = 'metadataService',
  DASHBOARD_SERVICE = 'dashboardService',
  PIPELINE_SERVICE = 'pipelineService',
  MLMODEL_SERVICE = 'mlmodelService',
  STORAGE_SERVICE = 'storageService',
  SEARCH_SERVICE = 'searchService',
  WEBHOOK = 'webhook',
  MLMODEL = 'mlmodel',
  TYPE = 'type',
  TEAM = 'team',
  USER = 'user',
  BOT = 'bot',
  ROLE = 'role',
  POLICY = 'policy',
  TEST_SUITE = 'testSuite',
  TEST_CASE = 'testCase',
  DATA_INSIGHT_CHART = 'dataInsightChart',
  KPI = 'kpi',
  ALERT = 'alert',
  CONTAINER = 'container',
  TAG = 'tag',
  DASHBOARD_DATA_MODEL = 'dashboardDataModel',
  SUBSCRIPTION = 'subscription',
  CHART = 'chart',
  DOMAIN = 'domain',
  DATA_PRODUCT = 'dataProduct',
  SAMPLE_DATA = 'sampleData',
  STORED_PROCEDURE = 'storedProcedure',
  SEARCH_INDEX = 'searchIndex',
  APP_MARKET_PLACE_DEFINITION = 'appMarketPlaceDefinition',
  APPLICATION = 'app',
  PERSONA = 'persona',
  DOC_STORE = 'docStore',
  PAGE = 'Page',
  knowledgePanels = 'KnowLedgePanels',
  GOVERN = 'govern',
  ALL = 'all',
  CUSTOM_METRIC = 'customMetric',
  INGESTION_PIPELINE = 'ingestionPipeline',
  QUERY = 'query',
  ENTITY_REPORT_DATA = 'entityReportData',
  WEB_ANALYTIC_ENTITY_VIEW_REPORT_DATA = 'webAnalyticEntityViewReportData',
  WEB_ANALYTIC_USER_ACTIVITY_REPORT_DATA = 'webAnalyticUserActivityReportData',
  TEST_CASE_RESOLUTION_STATUS = 'test_case_resolution_status_search_index',
  TEST_CASE_RESULT = 'test_case_result_search_index',
  EVENT_SUBSCRIPTION = 'eventsubscription',
  LINEAGE_EDGE = 'lineageEdge',
  API_SERVICE = 'apiService',
  API_COLLECTION = 'apiCollection',
  API_ENDPOINT = 'apiEndpoint',
  METRIC = 'metric',

  GEO = 'geo',
}